requirements/base.txt
django==5.0.6
asgiref==3.8.1
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-role-permissions==3.2.0
django-filter==24.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
django-redis==5.4.0
djangorestframework-simplejwt==5.3.1
drf-api-tracking==1.8.4
drf-yasg==1.21.7
django-environ==0.11.0
renku-sphinx-theme==0.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.6
setuptools==59.6.0
sphinx==7.3.7
zeep==4.2.1
